rename this code to be generic.  The Digest code here was written for RTSP, but is the same for http, so I'm going to use it for HTTP Digest auth.  This is just a rename.
